# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nexum REST API",
    "description": "REST API for the Nexum AI-powered development orchestration system. Provides CRUD operations for plans and tasks, status transitions, execution state inspection, agent claiming, and configuration access. All endpoints are versioned under `/api/v1/`. Write endpoints (POST, PUT, PATCH, DELETE) require Bearer token authentication via API key when authentication is enabled. Use GET /auth/status to discover authentication requirements.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
  "paths": {
...
    "/config": {
      "get": {
        "operationId": "getConfig",
        "summary": "Get current server configuration",
        "description": "Returns non-sensitive server configuration settings. API keys, OAuth tokens, and other secrets are excluded.",
        "tags": ["Config"],
        "responses": {
          "200": {
            "description": "Server configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/agents": {
      "get": {
        "operationId": "listAgents",
        "summary": "List registered agents",
        "description": "Returns all agent registrations from the configuration file.",
        "tags": ["Config"],
        "responses": {
          "200": {
            "description": "List of registered agents",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentsResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    }
  },
  "components": {
...
    "schemas": {
      "ApiErrorResponse": {
        "type": "object",
        "required": ["error", "message", "status"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Short error type name (e.g., 'NotFound', 'Validation', 'Conflict').",
            "enum": ["NotFound", "BadRequest", "Conflict", "Internal", "Unauthorized", "MethodNotAllowed", "Validation"]
          },
          "message": {
            "type": "string",
            "description": "Full human-readable error message."
          },
          "status": {
            "type": "integer",
            "format": "uint16",
            "description": "HTTP status code."
          }
        }
      },
...
      "ConfigResponse": {
        "type": "object",
        "required": ["server_host", "server_port", "max_parallel", "default_timeout_seconds", "log_level", "yolo_mode"],
        "properties": {
          "server_host": {
            "type": "string",
            "description": "Host the server binds to."
          },
          "server_port": {
            "type": "integer",
            "format": "uint16",
            "description": "Port the server listens on."
          },
          "max_parallel": {
            "type": "integer",
            "format": "uint16",
            "description": "Maximum number of tasks that may execute in parallel."
          },
          "default_timeout_seconds": {
            "type": "integer",
            "format": "uint64",
            "description": "Default timeout in seconds for task execution."
          },
          "log_level": {
            "type": "string",
            "description": "Log level (e.g., 'info', 'debug', 'warn')."
          },
          "yolo_mode": {
            "type": "boolean",
            "description": "Whether yolo mode is enabled (skips safety checks)."
          }
        }
      },
      "AgentRegistrationResponse": {
        "type": "object",
        "required": ["name", "type", "spawn_command"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Human-readable name of the agent."
          },
          "type": {
            "type": "string",
            "description": "Type/role of the agent."
          },
          "spawn_command": {
            "type": "string",
            "description": "Command used to spawn the agent process."
          },
          "tool_permissions": {
            "type": ["array", "null"],
            "items": {
              "type": "string"
            },
            "description": "Optional list of tool permissions."
          },
          "timeout_seconds": {
            "type": ["integer", "null"],
            "format": "uint64",
            "description": "Optional per-agent timeout in seconds."
          }
        }
      },
      "AgentsResponse": {
        "type": "object",
        "required": ["agents"],
        "properties": {
          "agents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentRegistrationResponse"
            },
            "description": "List of registered agents."
          }
        }
      },
...
    },
    "responses": {
...
      "InternalError": {
        "description": "Unexpected internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            },
            "example": {
              "error": "Internal",
              "message": "Internal error: IO operation failed",
              "status": 500
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key authentication. Include the API key in the Authorization header: 'Authorization: Bearer <api-key>'. Authentication is enabled via config and protects write endpoints by default."
      }
    },
...
  }
}